requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
chardet==3.0.4
charset-normalizer==3.2.0
frozenlist==1.3.3
html5lib==1.0.1
idna==2.10
lxml==4.9.3
mmh3==4.0.0
multidict==6.0.4
pip==23.1.2
python-nmap==0.7.1
requests==2.24.0
setuptools==68.0.0
six==1.16.0
soupsieve==2.4.1
urllib3==1.25.11
webencodings==0.5.1
wheel==0.40.0
yarl==1.9.2
